@@ -1,69 +1,41 @@
 {
-  "name": "kalkulator-dla-przedsiebiorcy-vue",
-  "version": "3.0.0",
-  "description": "Aplikacja Kalkulator finansowy jest bezpłatnym programem umożliwiającym łatwe obliczenie wynagrodzeń",
+  "name": "kalkulator-finansowy",
+  "version": "0.0.1",
+  "description": "Kalkulator finansowy jest bezpłatnym programem umożliwiającym łatwe obliczenie wynagrodzeń.",
   "productName": "Kalkulator finansowy",
-  "author": "Łuasz Socha <kontakt@lukasz-socha.pl>",
+  "author": "Łukasz Socha <kontakt@lukasz-socha.pl",
   "private": true,
   "scripts": {
-    "lint": "eslint --fix --ext .js,.vue,.ts ./",
-    "test": "echo \"See package.json => scripts for available tests.\" && exit 0",
-    "test:unit:ui": "majestic",
-    "test:unit": "jest --updateSnapshot",
-    "test:unit:ci": "jest --ci",
-    "test:unit:coverage": "jest --coverage",
-    "test:unit:watch": "jest --watch",
-    "test:unit:watchAll": "jest --watchAll",
-    "serve:test:coverage": "quasar serve test/jest/coverage/lcov-report/ --port 8788",
-    "concurrently:dev:jest": "concurrently \"quasar dev\" \"jest --watch\""
+    "lint": "eslint --ext .js,.ts,.vue ./",
+    "format": "prettier --write \"**/*.{js,ts,vue,scss,html,md,json}\" --ignore-path .gitignore",
+    "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "@quasar/cli": "^1.2.2",
-    "@quasar/extras": "^1.10.4",
-    "@typescript-eslint/eslint-plugin": "^5.5.0",
-    "@typescript-eslint/parser": "^5.5.0",
-    "@vue/compiler-sfc": "^3.0.11",
+    "@j-t-mcc/vue3-chartjs": "^1.3.0",
+    "@quasar/extras": "^1.0.0",
     "axios": "^0.21.1",
-    "chart.js": "^2.9.4",
-    "core-js": "^3.6.5",
     "date-fns": "^2.17.0",
-    "eslint-plugin-sort-keys-fix": "^1.1.2",
-    "quasar": "^2.0.4",
-    "vue3-chart-v2": "^0.8.2",
-    "vuex": "^4.0.2"
+    "pinia": "^2.0.11",
+    "quasar": "^2.7.4",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.0.0",
-    "@quasar/app": "^3.1.0",
-    "@quasar/quasar-app-extension-testing": "^2.0.2",
-    "@quasar/quasar-app-extension-testing-unit-jest": "^3.0.0-alpha.8",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-loader": "^4.0.2",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-jest": "^25.2.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.3.0",
-    "eslint-plugin-quasar": "^1.0.0",
+    "@quasar/app-vite": "^1.0.0",
+    "@types/node": "^12.20.21",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "autoprefixer": "^10.4.2",
+    "eslint": "^8.10.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
-    "eslint-plugin-vue": "^7.0.0",
-    "eslint-webpack-plugin": "^2.4.0",
-    "majestic": "^1.7.0",
-    "workbox-webpack-plugin": "^6.1.5"
+    "eslint-plugin-sort-keys-fix": "^1.1.2",
+    "eslint-plugin-vue": "^8.5.0",
+    "prettier": "^2.5.1",
+    "typescript": "^4.5.4"
   },
-  "browserslist": [
-    "last 10 Chrome versions",
-    "last 10 Firefox versions",
-    "last 4 Edge versions",
-    "last 7 Safari versions",
-    "last 8 Android versions",
-    "last 8 ChromeAndroid versions",
-    "last 8 FirefoxAndroid versions",
-    "last 10 iOS versions",
-    "last 5 Opera versions"
-  ],
   "engines": {
-    "node": ">= 10.18.1",
+    "node": "^18 || ^16 || ^14.19",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
