@@ -1,20 +1,20 @@
 <template>
   <div>
-    <div v-if="amount">
+    <div v-if="currencyConverterStore.amount">
       <div class="row justify-between q-px-md q-py-sm">
         <div>
-          {{ formatToCurrency(1, fromCurrency) }}
+          {{ formatToCurrency(1, currencyConverterStore.fromCurrency) }}
         </div>
         <div>
-          {{ formatToCurrency( valueForOne, toCurrency) }}
+          {{ formatToCurrency( currencyConverterStore.valueForOne, currencyConverterStore.toCurrency) }}
         </div>
       </div>
       <div class="row justify-between q-px-md q-py-sm bg-primary text-white">
         <div>
-          {{ formatToCurrency(amount, fromCurrency) }}
+          {{ formatToCurrency(currencyConverterStore.amount, currencyConverterStore.fromCurrency) }}
         </div>
         <div>
-          {{ formatToCurrency( valueForWholeAmount, toCurrency) }}
+          {{ formatToCurrency( currencyConverterStore.valueForWholeAmount, currencyConverterStore.toCurrency) }}
         </div>
       </div>
     </div>
@@ -26,21 +26,17 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
+import {defineComponent} from 'vue'
 import { formatToCurrency } from 'src/use/currencyFormat'
-import { mapGetters } from 'vuex'
-export default {
-  computed: {
-    ...mapGetters({
-      amount: 'currencyConverter/amount',
-      fromCurrency: 'currencyConverter/fromCurrency',
-      toCurrency: 'currencyConverter/toCurrency',
-      valueForOne: 'currencyConverter/valueForOne',
-      valueForWholeAmount: 'currencyConverter/valueForWholeAmount',
-    }),
-  },
+import {useCurrencyConverterStore} from 'stores/currency-converter-store'
+export default defineComponent({
   setup () {
-    return { formatToCurrency }
+    const currencyConverterStore = useCurrencyConverterStore()
+    return {
+      currencyConverterStore,
+      formatToCurrency,
+    }
   },
-}
+})
 </script>
