@@ -20,7 +20,7 @@
             label="Data*"
             :rules="[
               'date',
-              date => isPast(new Date(date)) || 'Wybierz datę z przeszłości'
+              date => isPast(new Date(currencyRateStore.date)) || 'Wybierz datę z przeszłości'
             ]">
             <template v-slot:append>
               <q-icon
@@ -46,16 +46,16 @@
     </q-form>
     <p
       class="text-center q-my-md text-primary">
-      <span v-if="date">
-        Kurs średni z dnia {{ date }}
+      <span v-if="currencyRateStore.date">
+        Kurs średni z dnia {{ currencyRateStore.date }}
       </span>
     </p>
     <q-table
-      :rows="rates"
+      :rows="currencyRateStore.currencyRates"
       :columns="columns"
       row-key="name"
       @row-click="openCurrency"
-      :loading="isLoading"
+      :loading="currencyRateStore.isLoading"
       hide-pagination
       :pagination="{rowsPerPage: 999}">
       <template v-slot:no-data>
@@ -71,17 +71,11 @@
 
 <script>
 import { format, isPast, isWeekend } from 'date-fns'
-import { mapGetters } from 'vuex'
+import {useCurrencyRateStore} from 'stores/currency-rate-store'
+import api from './api'
 import constants from 'src/logic/constants'
 
 export default {
-  computed: {
-    ...mapGetters({
-      date: 'exchangeRates/date',
-      isLoading: 'exchangeRates/isLoading',
-      rates: 'exchangeRates/rates',
-    }),
-  },
   created () {
     this.updateRates()
   },
@@ -125,24 +119,26 @@ export default {
       })
     },
     updateRates () {
+      const currencyRateStore = useCurrencyRateStore()
       let apiCall
 
-      this.$store.commit('exchangeRates/setLoading', true)
+      currencyRateStore.isLoading = true
 
       if (this.rateDate) {
         const rateDate = format(new Date(this.rateDate), 'Y-MM-dd')
-        apiCall = this.$store.dispatch('exchangeRates/loadExchangeRatesFromDate', { date: rateDate })
+        apiCall = api.loadExchangeRatesFromDate(rateDate)
       } else {
-        apiCall = this.$store.dispatch('exchangeRates/loadLatestExchangeRates')
+        apiCall = api.loadLatestExchangeRates()
       }
       apiCall.then(response => {
-        this.$store.commit('exchangeRates/setDate', response.data[0].effectiveDate)
-        this.$store.commit('exchangeRates/setRates', response.data[0].rates)
+        currencyRateStore.date = response.data[0].effectiveDate
+        currencyRateStore.currencyRates = response.data[0].rates
+
         this.$q.notify({
           message: 'Źródło danych: Narodowy Bank Polski',
         })
       }).catch((error) => {
-        this.$store.commit('exchangeRates/setDate', format(new Date(this.rateDate), 'Y-MM-dd'))
+        currencyRateStore.date = format(new Date(this.rateDate), 'Y-MM-dd')
         let message = 'Nie udało się połączyć z serwerem NBP. Spróbuj ponownie'
         if (error.response) {
           message = error.response.data
@@ -150,13 +146,14 @@ export default {
         this.$q.notify({
           message: message,
         })
-        this.$store.commit('exchangeRates/setRates', [])
+        currencyRateStore.currencyRates = []
       }).finally(() => {
-        this.$store.commit('exchangeRates/setLoading', false)
+        currencyRateStore.isLoading = false
       })
     },
   },
   setup () {
+    const currencyRateStore = useCurrencyRateStore()
     const onlyPast = (date) => {
       return date < format(new Date(), 'y/MM/dd')
     }
@@ -165,6 +162,7 @@ export default {
     }
     return {
       constants,
+      currencyRateStore,
       isPast,
       validDays (date) {
         return onlyPast(date) && onlyWorkDays(date)
