@@ -23,15 +23,18 @@
   </q-page>
 </template>
 
-<script>
-import Advert from 'components/partials/Advert'
-import CurrencyStatistics from 'components/exchangeRates/CurrencyStatistics'
-import Footer from 'components/partials/Footer'
-import Form from 'components/exchangeRates/Form'
-import SectionHeader from 'components/partials/SectionHeader'
+<script lang="ts">
+import {defineComponent} from 'vue'
+import {useAppStore} from 'stores/app-store'
+import {useCurrencyRateStore} from 'stores/currency-rate-store'
+import Advert from 'components/partials/Advert.vue'
+import CurrencyStatistics from 'components/exchangeRates/CurrencyStatistics.vue'
+import Footer from 'components/partials/Footer.vue'
+import Form from 'components/exchangeRates/Form.vue'
+import SectionHeader from 'components/partials/SectionHeader.vue'
 import helpers from 'src/logic/helpers'
 
-export default {
+export default defineComponent({
   components: {
     Advert,
     CurrencyStatistics,
@@ -40,13 +43,16 @@ export default {
     SectionHeader,
   },
   created () {
-    this.$store.commit('exchangeRates/setCurrency', null)
-    this.$store.commit('app/setModuleTitle', this.$route.params.currency.toUpperCase())
+    const appStore = useAppStore()
+    const currencyRateStore = useCurrencyRateStore()
+
+    appStore.moduleTitle = this.$route.params.currency.toUpperCase()
+    currencyRateStore.currencyRate = null
   },
   methods: {
     scrollTo () {
       helpers.scrollToElement(this.$refs.scrollTarget.$el)
     },
   },
-}
+})
 </script>
