@@ -15,21 +15,24 @@
   </q-page>
 </template>
 
-<script>
-import Advert from 'components/partials/Advert'
-import Footer from 'components/partials/Footer'
-import SectionHeader from 'components/partials/SectionHeader'
-import Table from 'components/exchangeRates/Table'
+<script lang="ts">
+import {defineComponent} from 'vue'
+import {useAppStore} from 'stores/app-store'
+import Advert from 'components/partials/Advert.vue'
+import Footer from 'components/partials/Footer.vue'
+import SectionHeader from 'components/partials/SectionHeader.vue'
+import Table from 'components/exchangeRates/Table.vue'
 
-export default {
+export default defineComponent({
   components: {
     Advert,
     Footer,
     SectionHeader,
     Table,
   },
-  created () {
-    this.$store.commit('app/setModuleTitle', 'Kursy walut')
+  setup () {
+    const appStore = useAppStore()
+    appStore.moduleTitle = 'Kursy walut'
   },
-}
+})
 </script>
