@@ -35,7 +35,7 @@
 <script lang="ts">
 import {UnregisteredCompanyInputFields} from 'components/unregisteredCompany/interfaces/UnregisteredCompanyInputFields'
 import {defineComponent, ref} from 'vue'
-import {useStore} from 'vuex'
+import {useAppStore} from 'stores/app-store'
 import Advert from 'components/partials/Advert.vue'
 import Footer from 'components/partials/Footer.vue'
 import Form from 'components/unregisteredCompany/Form.vue'
@@ -54,8 +54,8 @@ export default defineComponent({
     Summary,
   },
   setup() {
-    const store = useStore()
-    store.commit('app/setModuleTitle', 'Działalność niezarejestrowana')
+    const appStore = useAppStore()
+    appStore.moduleTitle = 'Działalność niezarejestrowana'
 
     const unregisteredCompanyInputFields = ref(<UnregisteredCompanyInputFields>{
       expenses: 0,
